phishing_project/requirements.txt
Flask==2.3.3
scikit-learn==1.2.2
pandas==2.2.2
numpy==1.26.4
joblib==1.3.2
requests==2.31.0
beautifulsoup4==4.12.2
tldextract==3.4.0
python-whois==0.7.3
cryptography==41.0.3
lxml==4.9.3
matplotlib==3.8.2
tqdm==4.66.1
